WIP
Improvements feature

+++ Assets/BusinessClicker/Scripts/Data/BusinessData.cs
@@ -10,20 +10,20 @@ namespace BusinessClicker.Data
         [SerializeField]
         private string _name;
         [SerializeField, Min(0.0f)]
-        private float _incomeDelay;
+        private double _incomeDelay;
         [SerializeField, Min(0.0f)]
-        private float _basePrice;
+        private double _basePrice;
         [SerializeField, Min(0.0f)]
-        private float _baseIncome;
+        private double _baseIncome;
 
         [Space, Header("Improvements")]
         [SerializeField]
         private BusinessImprovement[] _businessImprovements;
 
         public string Name => _name;
-        public float IncomeDelay => _incomeDelay;
-        public float BasePrice => _basePrice;
-        public float BaseIncome => _baseIncome;
+        public double IncomeDelay => _incomeDelay;
+        public double BasePrice => _basePrice;
+        public double BaseIncome => _baseIncome;
         public BusinessImprovement[] BusinessImprovements => _businessImprovements;
     }
 }
